+++ tests/test_django_rakaia/urls.py
@@ -8,5 +8,4 @@ from django.urls import include, path
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("streams/", include("django_rakaia.urls", namespace="django_rakaia")),
-    path("protocol/", include("django_rakaia.protocol_views", namespace="protocol")),
 ]